requirements.txt
absl-py==0.1.12
alembic==0.9.8
APScheduler==3.5.1
astor==0.6.2
bleach==1.5.0
chainer==3.4.0
chainercv==0.8.0
chainerui==0.2.0
click==6.7
cycler==0.10.0
filelock==3.0.4
fire==0.1.3
Flask==0.12.2
gast==0.2.0
grpcio==1.10.0
h5py==2.7.1
html5lib==0.9999999
itsdangerous==0.24
Jinja2==2.10
Keras==2.1.5
kiwisolver==1.0.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
matplotlib==2.2.2
numpy==1.14.2
Pillow==5.0.0
protobuf==3.5.1
pydot==1.2.4
pyparsing==2.2.0
python-dateutil==2.7.0
python-editor==1.0.3
pytz==2018.3
PyYAML==3.12
scipy==1.0.1
six==1.11.0
SQLAlchemy==1.2.5
tensorboard==1.7.0
tensorflow==1.7.0
tensorflowjs==0.1.1
termcolor==1.1.0
tzlocal==1.5.1
webdnn==1.2.2
Werkzeug==0.12.2
